Reorganize nextjs package
