+++ package.json
@@ -61,7 +61,7 @@
     "jest": "^29.0.3",
     "jest-environment-jsdom": "^29.0.3",
     "jest-location-mock": "^2.0.0",
-    "next": "^13.2.4",
+    "next": "^14.2.3",
     "node-fetch": "^2.7.0",
     "node-mocks-http": "^1.11.0",
     "ramda": "^0.29.0",
@@ -74,7 +74,7 @@
   },
   "dependencies": {
     "encoding": "^0.1.13",
-    "jose": "^4.6.0",
+    "jose": "^5.4.0",
     "path-to-regexp": "^6.2.1"
   },
   "jest": {}
