+++ src/index.ts
@@ -5,20 +5,22 @@
  */
 
 export * as browser from './browser';
+export type {
+  FiefAccessTokenInfo,
+  FiefParameters,
+  FiefTokenResponse,
+  FiefUserInfo,
+} from './client';
 export {
   Fief,
   FiefAccessTokenACRTooLow,
   FiefAccessTokenExpired,
-  FiefAccessTokenInfo,
   FiefAccessTokenInvalid,
   FiefAccessTokenMissingPermission,
   FiefAccessTokenMissingScope,
   FiefACR,
   FiefError,
   FiefIdTokenInvalid,
-  FiefParameters,
   FiefRequestError,
-  FiefTokenResponse,
-  FiefUserInfo,
 } from './client';
 export * as crypto from './crypto';
